# api-contract.yaml
openapi: 3.0.0
info:
  title: TarkVtark Debate Arena API
  description: |
    Backend API for TarkVtark debate application.
    This API manages debate topics, questions, replies, voting, and admin features.
    
    All data is stored in PostgreSQL database.
  version: 1.0.0
  contact:
    name: TarkVtark Team
    email: admin@tarkvtark.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.tarkvtark.com/api/v1
    description: Production server
...
paths:
...
  /topics:
    get:
      tags: [Topics]
      summary: Get all debate topics
      description: Returns a list of all active debate topics
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DebateTopic'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Topics]
      summary: Create a new debate topic
      description: Creates a new debate topic (admin only in production)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTopicRequest'
      responses:
        '201':
          description: Topic created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateTopic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /topics/{topicId}:
    get:
      tags: [Topics]
      summary: Get a specific topic by ID
      description: Returns a single topic with all its metadata
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebateTopic'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # =====================================================================
  # QUESTIONS ENDPOINTS
  # =====================================================================
  /questions/topic/{topicId}:
    get:
      tags: [Questions]
      summary: Get all questions for a topic
      description: Returns all questions with nested replies for a specific topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    DebateTopic:
      type: object
      required:
        - id
        - topic
        - leftLabel
        - rightLabel
        - isActive
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the topic
          example: "550e8400-e29b-41d4-a716-446655440000"
        topic:
          type: string
          description: Topic title (e.g., "Sanatan vs Islam")
          example: "Sanatan vs Islam"
          minLength: 5
          maxLength: 255
        leftLabel:
          type: string
          description: Label for the left column
          example: "Sanatan"
          minLength: 1
          maxLength: 100
        rightLabel:
          type: string
          description: Label for the right column
          example: "Islam"
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          description: Optional description of the debate topic
          example: "A respectful debate comparing philosophical aspects"
          maxLength: 1000
        isActive:
          type: boolean
          description: Whether the topic is currently active
          example: true
        createdAt:
          type: string
          format: date-time
          description: When the topic was created
          example: "2025-12-18T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the topic was last updated
          example: "2025-12-18T10:30:00Z"
    CreateTopicRequest:
      type: object
      required:
        - topic
        - leftLabel
        - rightLabel
      properties:
        topic:
          type: string
          minLength: 5
          maxLength: 255
          example: "Capitalism vs Socialism"
        leftLabel:
          type: string
          minLength: 1
          maxLength: 100
          example: "Capitalism"
        rightLabel:
          type: string
          minLength: 1
          maxLength: 100
          example: "Socialism"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "Economic systems debate"
        isActive:
          type: boolean
          default: true

    # -----------------------------------------------------------------
    # QUESTION
    # -----------------------------------------------------------------
    Question:
      type: object
      required:
        - id
        - debateTopicId
        - text
        - side
        - author
        - votesUp
        - votesDown
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the question
        debateTopicId:
          type: string
          format: uuid
          description: ID of the debate topic this question belongs to
        text:
          type: string
          description: The question text
          minLength: 10
          maxLength: 5000
        tag:
          type: string
          nullable: true
          description: Category tag (e.g., "Philosophy", "Theology")
          maxLength: 100
        side:
          type: string
          enum: [left, right]
          description: Which side of the debate (left or right column)
        author:
          type: string
          description: Username of the author
          default: "Anonymous"
          maxLength: 100
        votesUp:
          type: integer
          minimum: 0
          description: Number of upvotes
          example: 5
        votesDown:
          type: integer
          minimum: 0
          description: Number of downvotes
          example: 2
        uniqueId:
          type: string
          nullable: true
          description: Human-readable unique ID for sharing
          example: "q-1702475844123-456"
        replies:
          type: array
          description: Nested replies to this question
          items:
            $ref: '#/components/schemas/Reply'
        attachments:
          type: array
          description: File attachments for this question
          items:
            $ref: '#/components/schemas/Attachment'
        evidenceUrls:
          type: array
          description: Evidence URLs for this question
          items:
            $ref: '#/components/schemas/EvidenceUrl'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Reply:
      type: object
      required:
        - id
        - text
        - side
        - author
        - votesUp
        - votesDown
        - depth
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
          nullable: true
          description: ID of the question (if direct reply to question)
        parentReplyId:
          type: string
          format: uuid
          nullable: true
          description: ID of the parent reply (if reply to another reply)
        text:
          type: string
          minLength: 10
          maxLength: 5000
        side:
          type: string
          enum: [left, right]
          description: Always opposite of parent's side
        author:
          type: string
          default: "Anonymous"
          maxLength: 100
        votesUp:
          type: integer
          minimum: 0
        votesDown:
          type: integer
          minimum: 0
        uniqueId:
          type: string
          nullable: true
          example: "r-1702475844567-789"
        depth:
          type: integer
          minimum: 0
          description: Nesting level (0 = reply to question, 1 = reply to reply, etc.)
        replies:
          type: array
          description: Nested replies to this reply
          items:
            $ref: '#/components/schemas/Reply'
        attachments:
          type: array
          description: File attachments for this reply
          items:
            $ref: '#/components/schemas/Attachment'
        evidenceUrls:
          type: array
          description: Evidence URLs for this reply
          items:
            $ref: '#/components/schemas/EvidenceUrl'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Attachment:
      type: object
      required:
        - id
        - fileName
        - fileSize
        - fileType
        - storageUrl
        - storageProvider
        - uploadedBy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the attachment
        questionId:
          type: string
          format: uuid
          nullable: true
          description: ID of the question this is attached to
        replyId:
          type: string
          format: uuid
          nullable: true
          description: ID of the reply this is attached to
        fileName:
          type: string
          description: Original filename
          example: "evidence-document.pdf"
        fileSize:
          type: integer
          format: int64
          description: File size in bytes
          example: 2048576
        fileType:
          type: string
          description: MIME type
          example: "application/pdf"
        storageUrl:
          type: string
          format: uri
          description: URL to access the file
          example: "http://localhost:8080/api/v1/files/abc-123-456.pdf"
        storageProvider:
          type: string
          description: Storage provider (local, cloudinary, s3, etc.)
          example: "local"
          default: "local"
        displayOrder:
          type: integer
          nullable: true
          description: Order for display in UI
          example: 1
        uploadedBy:
          type: string
          description: Name of user who uploaded
          default: "Anonymous"
          maxLength: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EvidenceUrl:
      type: object
      required:
        - id
        - url
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the evidence URL
        questionId:
          type: string
          format: uuid
          nullable: true
          description: ID of the question this is linked to
        replyId:
          type: string
          format: uuid
          nullable: true
          description: ID of the reply this is linked to
        url:
          type: string
          format: uri
          description: The evidence URL
          example: "https://youtube.com/watch?v=example"
        title:
          type: string
          nullable: true
          description: Optional title/description
          maxLength: 255
          example: "Supporting Evidence Video"
        displayOrder:
          type: integer
          nullable: true
          description: Order for display in UI
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # -----------------------------------------------------------------
    # ERROR RESPONSES
    # -----------------------------------------------------------------
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details

  # =====================================================================
  # REUSABLE RESPONSES
  # =====================================================================
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "BAD_REQUEST"
            message: "Invalid input data"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_ERROR"
            message: "An unexpected error occurred"